requirements.txt
Django==2.2
djangorestframework==3.12.2
drf-writable-nested==0.6.2
pkg-resources==0.0.0
pytz==2020.4
sqlparse==0.4.1
